force use of unsigned java servlet API
By default Spark uses the eclipse orbit version of java servlet API, which is signed.  This leads to security exceptions when spark test cases are run from IDEA because the implementation of java servlet that gets picked up is unsigned.  This is alleviated by forcing the use of the javax version of servlet api, which is unsigned as well.