Backs out negative-handling log transformer, replaces with a basic log10
Requires a kludge in the form of fixing the minimum range to 1 in cases that the data contains zeros. If data is expected to have an "illegal" minimum but the range between 0 and 1 is important to show the default min will have to be configured to something else that makes sense for the data, say 0.001.